refactor: extract AES-256-CBC crypto into shared utility
